PRJCT/python-core/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.13.0
cycler==0.12.1
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.128.0
feedparser==6.0.12
filelock==3.20.3
fonttools==4.61.1
frozendict==2.4.7
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
idna==3.11
ib_insync==0.9.86
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
multitasking==0.0.12
numpy==2.2.6
packaging==25.0
pandas==2.3.3
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymongo==4.16.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
yfinance==1.0
